requirements.txt
# Core FastAPI and server
fastapi==0.109.2 # verify
uvicorn==0.27.0 # verify

# Database and ORM
sqlalchemy==2.0.25 # verify
pydantic==2.6.3 # verify

# Password hashing
passlib[bcrypt]==1.7.4 # passlib is not updated frequently, verify if needed

# Environment variables
python-dotenv==1.0.1 # verify

# JWT for session management
python-jose[cryptography]==3.3.0 # Verify, may have newer versions

# Redis client
redis==5.0.3 # verify

# Celery and RabbitMQ
celery==5.3.6 # verify celery, check if there are newer versions
pika==1.3.2 # check if there are newer versions
flower==2.0.1 #monitor celery


#email
google-auth-oauthlib==1.2.0
google-api-python-client==2.125.0
# Testing
pytest==8.0.2 # verify
pytest-asyncio==0.23.3 # verify
httpx==0.26.0 # verify
requests==2.31.0

